Break up deps tests to be a little less intimidating.
